[WIP] Fix errors after dir struct update
After rebasing the code with main a lot of issues arrised.
This is just a workaround to make it possible to compile and run
it again, but there are a lot of refactor still needed.
