+++ controllers/backup/cr.go
@@ -25,7 +25,7 @@ func (r *PulpBackupReconciler) backupCR(ctx context.Context, pulpBackup *repoman
 	execCmd := []string{
 		"bash", "-c", "echo '" + string(pulpSpec) + "' > " + backupDir + "/cr_object",
 	}
-	_, err := containerExec(pod, r, execCmd, pulpBackup.Name+"-backup-manager", pod.Namespace)
+	_, err := r.containerExec(pod, execCmd, pulpBackup.Name+"-backup-manager", pod.Namespace)
 	if err != nil {
 		log.Error(err, "Failed to backup "+pulpBackup.Spec.DeploymentType+" CR")
 		return err
